backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
django-seed==0.3.1
djangorestframework==3.15.2
Faker==33.0.0
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
jalali_core==1.0.0
jdatetime==5.0.0
lxml==5.3.0
multidict==6.1.0
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.5.2
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.0
zeep==4.3.1
